pyproject.toml
[tool.poetry]
name = "db-replay"
version = "0.1.0"
description = ""
authors = ["Filip Sedlák <filip@sedlakovi.org>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.2"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
